report: move types around
